refactor(system_configurator): split into package with separate modules
Split the monolithic system_configurator.py into a package for better
organization as more configurators are added. base.py has the ABC and
helpers, clock_sync.py has ClockSyncConfigurator, lcm.py has LCM
configurators. __init__.py re-exports everything for backward compat.
