+++ dimos/protocol/service/system_configurator/__init__.py
@@ -1,0 +1,51 @@
+# Copyright 2026 Dimensional Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""System configurator package — re-exports for backward compatibility."""
+
+from dimos.protocol.service.system_configurator.base import (
+    SystemConfigurator,
+    _is_root_user,
+    _read_sysctl_int,
+    _write_sysctl_int,
+    configure_system,
+    sudo_run,
+    system_checks,
+)
+from dimos.protocol.service.system_configurator.clock_sync import ClockSyncConfigurator
+from dimos.protocol.service.system_configurator.lcm import (
+    IDEAL_RMEM_SIZE,
+    BufferConfiguratorLinux,
+    BufferConfiguratorMacOS,
+    MaxFileConfiguratorMacOS,
+    MulticastConfiguratorLinux,
+    MulticastConfiguratorMacOS,
+)
+
+__all__ = [
+    "IDEAL_RMEM_SIZE",
+    "BufferConfiguratorLinux",
+    "BufferConfiguratorMacOS",
+    "ClockSyncConfigurator",
+    "MaxFileConfiguratorMacOS",
+    "MulticastConfiguratorLinux",
+    "MulticastConfiguratorMacOS",
+    "SystemConfigurator",
+    "_is_root_user",
+    "_read_sysctl_int",
+    "_write_sysctl_int",
+    "configure_system",
+    "sudo_run",
+    "system_checks",
+]
