+++ dimos/protocol/service/system_configurator/lcm.py
@@ -14,278 +14,16 @@
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from functools import cache
-import os
-import platform
 import re
 import resource
-import socket
-import struct
 import subprocess
-import time
-from typing import TYPE_CHECKING, Any
 
-if TYPE_CHECKING:
-    from collections.abc import Callable
-
-# ----------------------------- sudo helpers -----------------------------
-
-
-@cache
-def _is_root_user() -> bool:
-    try:
-        return os.geteuid() == 0
-    except AttributeError:
-        return False
-
-
-def sudo_run(*args: Any, **kwargs: Any) -> subprocess.CompletedProcess[str]:
-    if _is_root_user():
-        return subprocess.run(list(args), **kwargs)
-    return subprocess.run(["sudo", *args], **kwargs)
-
-
-def _read_sysctl_int(name: str) -> int | None:
-    try:
-        result = subprocess.run(["sysctl", name], capture_output=True, text=True)
-        if result.returncode != 0:
-            print(
-                f"[sysctl] ERROR: `sysctl {name}` rc={result.returncode} stderr={result.stderr!r}"
-            )
-            return None
-
-        text = result.stdout.strip().replace(":", "=")
-        if "=" not in text:
-            print(f"[sysctl] ERROR: unexpected output for {name}: {text!r}")
-            return None
-
-        return int(text.split("=", 1)[1].strip())
-    except Exception as error:
-        print(f"[sysctl] ERROR: reading {name}: {error}")
-        return None
-
-
-def _write_sysctl_int(name: str, value: int) -> None:
-    sudo_run("sysctl", "-w", f"{name}={value}", check=True, text=True, capture_output=False)
-
-
-# -------------------------- base class for system config checks/requirements --------------------------
-
-
-class SystemConfigurator(ABC):
-    critical: bool = False
-
-    @abstractmethod
-    def check(self) -> bool:
-        """Return True if configured. Log errors and return False on uncertainty."""
-        raise NotImplementedError
-
-    @abstractmethod
-    def explanation(self) -> str | None:
-        """
-        Return a human-readable summary of what would be done (sudo commands) if not configured.
-        Return None when no changes are needed.
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    def fix(self) -> None:
-        """Apply fixes (may attempt sudo, catch, and apply fallback measures if needed)."""
-        raise NotImplementedError
-
-
-# ----------------------------- generic enforcement of system configs -----------------------------
-
-
-def configure_system(checks: list[SystemConfigurator], check_only: bool = False) -> None:
-    if os.environ.get("CI"):
-        print("CI environment detected: skipping system configuration.")
-        return
-
-    # run checks
-    failing = [check for check in checks if not check.check()]
-    if not failing:
-        return
-
-    # ask for permission to modify system
-    explanations: list[str] = [msg for check in failing if (msg := check.explanation()) is not None]
-
-    if explanations:
-        print("System configuration changes are recommended/required:\n")
-        print("\n\n".join(explanations))
-        print()
-
-    if check_only:
-        return
-
-    try:
-        answer = input("Apply these changes now? [y/N]: ").strip().lower()
-    except (EOFError, KeyboardInterrupt):
-        answer = ""
-
-    if answer not in ("y", "yes"):
-        if any(check.critical for check in failing):
-            raise SystemExit(1)
-        return
-
-    for check in failing:
-        try:
-            check.fix()
-        except subprocess.CalledProcessError as error:
-            if check.critical:
-                print(f"Critical fix failed rc={error.returncode}")
-                print(f"stdout: {error.stdout}")
-                print(f"stderr: {error.stderr}")
-                raise
-            print(f"Optional improvement failed: rc={error.returncode}")
-            print(f"stdout: {error.stdout}")
-            print(f"stderr: {error.stderr}")
-
-    print("System configuration completed.")
-
-
-# ----------------------------- bridge: SystemConfigurator → Blueprint.requirements() -----------------------------
-
-
-def system_checks(*configurators: SystemConfigurator) -> Callable[[], str | None]:
-    """Wrap SystemConfigurator instances into a Blueprint.requirements()-compatible callable.
-
-    Returns a function that runs configure_system() and converts SystemExit
-    (raised when a critical check is declined) into an error string.
-    Non-critical declines return None (proceed with degraded state).
-    """
-
-    def _check() -> str | None:
-        try:
-            configure_system(list(configurators))
-        except SystemExit:
-            labels = [type(c).__name__ for c in configurators]
-            return f"Required system configuration was declined: {', '.join(labels)}"
-        return None
-
-    return _check
-
-
-# ------------------------------ specific checks: clock sync ------------------------------
-
-
-class ClockSyncConfigurator(SystemConfigurator):
-    """Check that the local clock is within MAX_OFFSET_SECONDS of NTP time.
-
-    Uses a pure-Python NTP query (RFC 4330 SNTPv4) so there are no external
-    dependencies.  If the NTP server is unreachable the check *passes* — we
-    don't want unrelated network issues to block robot startup.
-    """
-
-    critical = False
-    MAX_OFFSET_SECONDS = 0.1  # 100 ms per issue spec
-    NTP_SERVER = "pool.ntp.org"
-    NTP_PORT = 123
-    NTP_TIMEOUT = 2  # seconds
-
-    def __init__(self) -> None:
-        self._offset: float | None = None  # seconds, filled by check()
-
-    # ---- NTP query ----
-
-    @staticmethod
-    def _ntp_offset(server: str = "pool.ntp.org", port: int = 123, timeout: float = 2) -> float:
-        """Return clock offset in seconds (local - NTP).  Raises on failure."""
-        # Minimal SNTPv4 request: LI=0, VN=4, Mode=3 → first byte = 0x23
-        msg = b"\x23" + b"\x00" * 47
-        sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-        sock.settimeout(timeout)
-        try:
-            t1 = time.time()
-            sock.sendto(msg, (server, port))
-            data, _ = sock.recvfrom(1024)
-            t4 = time.time()
-        finally:
-            sock.close()
-
-        if len(data) < 48:
-            raise ValueError(f"NTP response too short ({len(data)} bytes)")
-
-        # Transmit Timestamp starts at byte 40 (seconds at 40, fraction at 44)
-        ntp_secs: int = struct.unpack("!I", data[40:44])[0]
-        ntp_frac: int = struct.unpack("!I", data[44:48])[0]
-        # NTP epoch is 1900-01-01; Unix epoch is 1970-01-01
-        ntp_time: float = ntp_secs - 2208988800 + ntp_frac / (2**32)
-
-        # Simplified offset: assume symmetric delay
-        t_server = ntp_time
-        rtt = t4 - t1
-        offset: float = t_server - (t1 + rtt / 2)
-        return offset
-
-    # ---- SystemConfigurator interface ----
-
-    def check(self) -> bool:
-        try:
-            self._offset = self._ntp_offset(self.NTP_SERVER, self.NTP_PORT, self.NTP_TIMEOUT)
-        except (TimeoutError, OSError, ValueError) as exc:
-            print(f"[clock-sync] NTP query failed ({exc}); assuming clock is OK")
-            self._offset = None
-            return True  # graceful degradation — don't block on network issues
-
-        if abs(self._offset) <= self.MAX_OFFSET_SECONDS:
-            return True
-
-        print(
-            f"[clock-sync] WARNING: clock offset is {self._offset * 1000:+.1f} ms "
-            f"(threshold: ±{self.MAX_OFFSET_SECONDS * 1000:.0f} ms)"
-        )
-        return False
-
-    def explanation(self) -> str | None:
-        if self._offset is None:
-            return None
-        offset_ms = self._offset * 1000
-        system = platform.system()
-        if system == "Linux":
-            cmd = "sudo timedatectl set-ntp true && sudo systemctl restart systemd-timesyncd"
-        elif system == "Darwin":
-            cmd = "sudo sntp -sS pool.ntp.org"
-        else:
-            cmd = "(manual NTP sync required for your platform)"
-        return (
-            f"- Clock sync: local clock is off by {offset_ms:+.1f} ms "
-            f"(threshold: ±{self.MAX_OFFSET_SECONDS * 1000:.0f} ms)\n"
-            f"  Fix: {cmd}"
-        )
-
-    def fix(self) -> None:
-        system = platform.system()
-        if system == "Linux":
-            sudo_run(
-                "timedatectl",
-                "set-ntp",
-                "true",
-                check=True,
-                text=True,
-                capture_output=True,
-            )
-            sudo_run(
-                "systemctl",
-                "restart",
-                "systemd-timesyncd",
-                check=True,
-                text=True,
-                capture_output=True,
-            )
-        elif system == "Darwin":
-            sudo_run(
-                "sntp",
-                "-sS",
-                self.NTP_SERVER,
-                check=True,
-                text=True,
-                capture_output=True,
-            )
-        else:
-            print(f"[clock-sync] No automatic fix available for {system}")
-
+from dimos.protocol.service.system_configurator.base import (
+    SystemConfigurator,
+    _read_sysctl_int,
+    _write_sysctl_int,
+    sudo_run,
+)
 
 # ------------------------------ specific checks: multicast ------------------------------
